refactor: 테스트 코드를 CarTest 에서 CarNameTest 로 따로 분리
