+++ src/test/java/racingcar/domain/CarTest.java
@@ -58,53 +58,4 @@ public class CarTest {
 		}
 	}
 
-	@Nested
-	@DisplayName("자동차 이름에 ")
-	class CarNameTest {
-
-		@ParameterizedTest
-		@ValueSource(strings = {",", "#", "#phiz", "ph,bi"})
-		@DisplayName("허용되지 않는 문자를 입력했을 때 예외가 발생한다")
-		public void Not_Available_Character(String carName) {
-			assertThatThrownBy(
-				() -> Car.builder()
-					.name(carName)
-					.build())
-				.isInstanceOf(RuntimeException.class)
-				.hasMessage(NOT_ALLOWED_FORMAT_MESSAGE);
-
-		}
-
-		@Test
-		@DisplayName("null을 입력했을 때 예외가 발생한다")
-		public void Not_Available_Null() {
-			assertThatThrownBy(
-				() -> Car.builder()
-					.name(null)
-					.build())
-				.isInstanceOf(RuntimeException.class)
-				.hasMessage(INVALID_CAR_NAME_SHOULD_BE_ONE_MORE_CHARACTER);
-		}
-
-		@Test
-		@DisplayName("차 이름이 5글자를 넘어갔을 경우 예외가 발생한다")
-		public void input_car_name_exceed_length() {
-			assertThatThrownBy(
-				() -> Car.builder()
-					.name("abcdef")
-					.build())
-				.isInstanceOf(RuntimeException.class)
-				.hasMessage(EXCEED_LENGTH_ERROR_MESSAGE);
-		}
-
-		@ParameterizedTest
-		@ValueSource(strings = {"a", "aa", "philz"})
-		@DisplayName("정상 입력을 했을 때 예외가 발생하지 않는다")
-		public void Available_Character(String carName) {
-			Assertions.assertDoesNotThrow(
-				() -> Car.builder()
-					.name(carName)
-					.build());
-		}
-	}
 }
